+++ itk_dev_shared_components/eflyt/eflyt_util.py
@@ -1,0 +1,7 @@
+"""General helper funtions"""
+from datetime import date
+
+
+def format_date(_date: date) -> str:
+    """Format date as %d-%m-%Y"""
+    return _date.strftime("%d-%m-%Y")