+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - misc.file_util: handle_save_dialog
+- cvr_api.cvr_lookup: Look up cvr number.
 
 ### Changed