+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added modules for use with the Eflyt / Dedalus / Notus address system
 - misc.file_util: handle_save_dialog
+- cvr_api.cvr_lookup: Look up cvr number.
 
 ### Changed